Revamped website with many changes.
Website has a revamped navigation bar, with new hover effects that increase the user experience and interactivity. The home page is now complete, with a font style that matches the website's feel. The releases page now contains the band's newest album, with relevant information about the album. The other pages function, but do not contain information.