Completed contact page and minor edits
Began and finished the contact page of my website, containing a form that can be edited with the user's input. However, it does not actually send the message to the desired email. The real website will have a working message system. I also fixed some bugs about root element re-sizing on the releases page.